Simple script to enable remote console
Basically this sets up a listen1 with an rc (how we have done it on many machines
forever).

It's utterly insecure (hence port 1522 -- issh, i.e. is22). But it's incredibly handy
until we get ssl or something.

Known problem is you have to put  at the end of commands, since telnet is sending some
newline thing that we don't understand. So type
ls<space><return>
not ls<return>
until we fix this.

I don't know where to put this, somebody tell me. It really needs to go in a startup
script, after ipconfig, I guess.

Change-Id: I8ce690f072fac87fcbdaceab9a1c0dbb7de9997a
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>