Fixes website's top menu jumpiness when screen < 1000px wide.
The website's menu is quite wide and the default Bootstrap3 breakpoints don't play well with it at smaller screen sizes (ie. it shows a wonky menu spanning two lines).

This fix alters the default behaviour for the 768px-992px range so devices with these screen sizes get the mobile menu.

Change-Id: Ic43bce1e264b62ca1df754e1144b0d0ec68893e9