Fixes website's top menu jumpiness when screen < 1000px wide.
The website's menu is quite wide and the default Bootstrap3 breakpoints don't play well with it at smaller screen sizes (ie. it shows a wonky menu spanning two lines).

This fix alters the default behaviour for the 768px-992px range so devices with these screen sizes get the mobile menu.

Change-Id: Iaf5122980fa6f85f3878e389bb30909a7b04e974