Add the regression device from NxM
This device allows you to exercise known bad code in the kernel.

Let's say you find a kernel bug in, e.g., tsleep. You write a test in
devregress.c that allows you to reproduce the bug, and you write a program
(put source in sys/src/regress) that opens devregress and commands it to
take the action to exercise the bug. Then, you fix the bug, run the program
again, and the bug should not occur.

If you're going to fix (e.g.) tsleep, be sure to write the regression and reproducer
first, so we know it stays fixed.

Change-Id: Ic021f39217081a586ebbfe0c81f2d9f4f7eec0c0
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>