Preen files. Walk the tree, remove cflags, put things in order.
This is not tested in the sense that I don't know if we end up building working
code :-)

But it did show to me the value of having build files using JSON that can be automatically processed
by a Go tool. We had gotten a little sloppy with our json files, in that things were getting
out of alphabetical order, not quite formatted right, and so on.

There are other issues, such as the use of ../../../ where  a /sys/src/cmd
will do better. All that can be fixed with this tool, and we can make sure .json files
follow the rules.

Note that we can ignore files by setting up the Except array. We currently ignore /cflags.json and the entire 9/ directory.

Next steps are to create /amd64/include/cflags.json (or something; /include/cflags.json?) and actually use it. This can
remove the duplicate definition of Cflags everywhere.

This tool can grow over time to be a general purpose scanner for our .json files to make sure they are not bit rotting. There's
more to do but this is a start. I want to get it working but I'm hoping someone else will also be interested. We also need to figure
out what other things we want to fix. I have a list.

Also, this is a general pattern we should follow: never do things by hand. Harvey should learn from the Bell Labs traditions, one
of which includes automating things with small, smart programs.

Change-Id: Icc88e83f58855a5805de21923f85650433ed4b10
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>