Add an alternate mount device, 'N', a.k.a. mntn
This shows how to add a new mount device. I leave changing the mount command
to take a parameter selecting this device as an exercise for the reader
(hint: it's easy).

Note that we share lots of code with devmnt. This might not be the case for another
mount device. This is a proof of concept.

Change-Id: I40a2f54338617e183e59473f16c51aaf220e72e8
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>