Add the option ROM support from coreboot
This is the essential part of doing option ROM support.

I think we'll want to remove the realmode directory, as that requires
dropping all protections and going to 16-bit x86 mode, rarely a good idea;
and, besides, it's just gross.

I think our best bet is to use yabel, as coreboot has for many years.
I'm putting it all here as an initial commit so we can pick over its bones.

Change-Id: I61a979ac4808dd6f1f278b8a8d908749768bba52
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>