x86emu: getting it close to compiling.
It now builds but does not link. Just oding one file so far.
Getting rid of as much "include includes includes" and other such
brain damage as I can too. It's amazing how many times these packages
include the same files over and over. No wonder it's all so slow to compile!

This builds. Now to get the rest of the files in. I could use the help.

Change-Id: I81493ebfb655578002c20a4ac729701682224045
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>