libc: fixed signature of getcallerpc(2), ref #19
The implementation was already based on [__builtin_return_address](https://gcc.gnu.org/onlinedocs/gcc/Return-Address.html) (which is supported by both GCC and clang), so there is no need for the first argument's address.

In theory, we should call `__builtin_extract_return_addr` on the address provided by `__builtin_return_address`, but on AMD64 this is an identity function, so we ignore it for now.

Change-Id: Ia4d8c565de8feb16ee9ed17f21f361c6a5632f28
Signed-off-by: Giacomo Tesio <giacomo@tesio.it>