Change consolefs
We may want to change it back but it was never all that useful, as
witness the fact none of theo old guard on this list even remembered it.

This is a hack. It now reads 0 and writes 1 and presents /dev/cons and /dev/consctl.

Not sure at present why it doesn't want to do consctl but this is a start.

Next steps: rip the keyboard and display handling out of console.c and put it in here as an option:
[0,1] can be [open(keyboard), open(cga)]

Change-Id: I7169e2752d36792ddb80b44729dd87bc2d06903f
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>