DRAFT: do not commit: temporarily revive kernel console device
This is a TEMPORARY HACK so that we can have a working /dev/cons.

Fails badly, try it out. But it might work with a tiny bit of tweaking.

This will be replaced, we hope soon, by an improved sys/src/cmd/aux/console.c
which exports a cons/consctl pair that can be mounted into /dev, and an improved
sys/src/cmd/aux/tty.c that opens /dev/cons for IO, instead of just using stdin/out

This is still not working and will in fact lock up your console, probably as it competes with
the user mode console. Sorry, folks, just me fooling around with not enough time to do it right.
Maybe somebody can take this and make it go.

Change-Id: I615810a08b5fd1668ef668c0f9d9a2161a8583b5
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>