Add aux/nconsole.c 9p server, which implements /dev/cons and /dev/consctl
For now, re enable debug prints. Also, chatty9p is bumped, so you'll get
messages about 9p traffic.

Once the mount is done, it forks and the child opens /dev/cons and dups
it over [0,1,2], and execs the args as console does now.

In theory, this should print the date:
nconsole /bin/date

But it does not work.

Typically, nconsole would fork rc, and then we would be able to run
commands that need to open /dev/cons

Change-Id: I124ced703b5cffa192e9855ee8b421b9683a99b0
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>