Add aux/nconsole.c 9p server to replace console.c
This is just like console.c now, save that it implements /dev/cons and /dev/consctl

To enable debug prints use -d, and chatty9p is also bumped, so you'll get
messages about 9p traffic.

To enable stdio instead of the ps2 keyboard and cga, use -s for stdinmode.
This is the case where you launch from listen1.

Once the mount is done, it forks and the child opens /dev/cons and dups
it over [0,1,2], and execs the args as console does now.

In theory, this should print the date:
nconsole /bin/date

And it does, :-), but nconsole does not exit at the end :-(

Typically, nconsole would fork rc, and then we would be able to run
commands that need to open /dev/cons

This is very close, you can even see the error print after exec works
on fd2, which is /dev/cons from the server. Why the exec fails is mystery.

Exec now works. The final bit is to close on the last close of cons.

So close ... someone take a look. I'm done for today.

Change-Id: I124ced703b5cffa192e9855ee8b421b9683a99b0
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>