Simple error benchmark
This times how many bad system calls we can make in
a second. This will be useful for measuring setjmp/longjmp
improvements.

The eventual way we should do this:
in bench, create a directory with the ref for tip
mkdir `git show-ref HEAD --hash`
run each benchmark and put output in a file for
results/ref/benchname

Then, over time, we accumulate results we can tie to a certain ref.
Example included.

Change-Id: Ie4f00e53fde7c99861a655280bfcc06ef516a1ec
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>