Build: rework it wo we can pass a directory and assume build.json
This has some problem but I can do no more today.

Basically, it build will work with directories and, if given a directory,
will assume the file is build.json. Or, if given a .json, will use it.

Also added a debug flag. It fails with BUILD all but I don't know why yet.

Change-Id: I0702583c4a3e3e8537e2681a582a42d210d0a1be
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>