Build: rework it so we can pass a directory and assume build.json
We can now simply name a directory if the .json file is build.json

If given a directory, build will assume the file is build.json.

Or, if given a .json file, will use it by changing to that directory and
using the file.

This applies to .json scripts with embedded names. You can name the
directory only if the file in that directory is build.json.

Also added a debug flag.

This works with a full BUILD and boot

Change-Id: I0702583c4a3e3e8537e2681a582a42d210d0a1be
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>