JSON: build, preen, mksys: make the .json files be arrays of json structs
This is a first step to allowing us to bundle things in json files as arrays,
which will in turn let us select individual elements by name or re matching
the Name.

This will in turn let us remove the BUILD script, since we'll have multiple
elements in the .json file and we can pick them by name.

Note that this change was done by the preen tool, save for core.json,
which I had to do by hand. Preen is handy. It lets you clean things up.

Next step is to have preen break SourceFilesCmd clauses into individual
JSON entities with names. Then we can remove the hacky SourceFilesCmd
part of build entirely. It was never a great thing anyway -- just an
expedient hack.

Change-Id: Iee0c6e9738a785a1e064291e943d8498c2321f41
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>