New json format per Giacomo's ideas.
Instead of this:
{
	"Name": "a",
	...
}

Do this:
{
	"Name": {
		...
	}
}

This is way better. Now the name is visible at the top level of any json viewer, we can
have multiple stanzas per file, the name is way more easy to find, ... the advantages are legion.

This includes a change to build.go but not to preen.

kernel building fails, because it includes all the files twice. Someone want to look? I'm done for the morning,
time for my day job.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

Change-Id: I9b1cf2597e582895b8f3ab127bc9c1d77de96cbd