New json format per Giacomo's ideas.
Instead of this:
{
	"Name": "a",
	...
}

Do this:
{
	"Name": {
		...
	}
}

This is way better. Now the name is visible at the top level of any json viewer, we can
have multiple stanzas per file, the name is way more easy to find, ... the advantages are legion.

So if this format meets approval I'll next change build and preen to go with it.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

Change-Id: I9b1cf2597e582895b8f3ab127bc9c1d77de96cbd