Break sys/src/cmds/cmds.json into two pieces
One piece is Dirs, the other Cmds.

If you type
build sys/src/cmds/cmds.json
the default pattern is .*, and both will be built.
If you type
build sys/src/cmds/cmds.json Cmds (or C.* for that matter

then only the single-file commands will be built, which is a time saver.

Change-Id: I0a782c36bd48bdb6cd13cfea653ac91db2b2a1fd
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>