Add the famed mclock.
This required *NO* changes for harvey. It's a model of how to write
portable graphics code.

There's an interesting back story here. In 1975 or so, at University of
Delaware, there was competition between the advocates of the PLATO
system, (wikipedia plato cdc), which ran on very expensive
supercomputers, and the RSTS/E system, that ran on a 100x cheaper PDP 11
minicomputer.

Roughly, looking at relative money and power, the PLATO team were the
suits, and the minicomputer guys were the rebel alliance. At least
that's how it seemed to us at the time ...

Much was made of the PLATO system's graphics, which were a big deal at
the time. My then advisor, Dr. David M. Robinson, believed the
minicomputer could equal the interactive performance of the CDC Cyber
mainframe. At the time we had the also-then-new Tektronix graphics
terminals, which used a storage tube technology. Dr. Robinson asked one
of the high school teenage programmers to write a simple clock example,
in Basic, and the result was the ancestor of this program.

In 1977, John DeGood rewrote it in HP 3000 Fortran for the HP 2648A
raster scan graphics terminal. He added color when HP introduced a color
graphics terminal, the HP 2627A, ca. 1982.

Fast forward 3 decades. John wrote a new Plan 9 version for the 2007
International Workshop on Plan 9 at Murray Hill Bell Labs (where it all
began for Unix), based on the original UDel "green bar" (search green
bar line printer paper :image) listing Dave Robinson gave him in 1977.
Yes, somehow the listing survived 30 years and 3 moves! John found it in
his basement. It has to have been the only copy, and the media was lost
long ago.

John hand-converted the regions to the arrays you see in this program (!),
learned Plan 9 graphics programming, and here we are.

David Robinson was a wonderful professor and a real visionary in pushing
computing to everyone. It's nice that he's represented in Harvey, since
Dave also worked at Bell Labs before coming to udel. He also had a great
sense of humor and I think he'd get a good laugh at seeing this program
live on.

Change-Id: Iba32b2569d7a583215a2c241803fddd3b71c9a47
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>