kernel: move sleep(2) to libc, see #29 (and #25)
`void sleep(int64_t millisecs)`

Sleep suspends the current process for the number of milliseconds
specified by the argument.  The actual suspension time may be a
little more or less than the requested time.
If millisecs is 0, the process gives up the CPU if another
process is waiting to run, returning immediately if not.
Implemented in terms of awake(2) and rendezvous(2),
it cannot be interrupted.

A few clients required little updates:

- in libthread, the (undocumented) iosleep() function
- in sys/src/cmd/auth/cron.c, the sleepuntil() function
- in sys/src/cmd/aux/apm.c, the eventwatch() function
- in sys/src/cmd/pump.c, the sleepunlocked() function

Whenever the return value of a function was depending on the return value
of the sleep() syscall, 0 (zero) is returned instead.

A new test `regress/sleep` has been introduced, and `regress/awake`
and `regress/alarm` have been updated to match the new signature and semantic.

	regress/sleep	PASS
	regress/awake	PASS
	regress/alarm	PASS

Change-Id: Id431b8e6f4ad2b6d39ed50e34cc81453bc38a8cb
Signed-off-by: Giacomo Tesio <giacomo@tesio.it>