usb: mv scsierrs.c to libdisk
This is the start of work to make usb more like a real
system instead of a bag on the side.

It will require moving .h files, creating a libusb, and separating
commands from libraries.

First step is to move things out of usb that are more general.

scsierrs.c is not only usb, it's everything scsi.

We can eventually consider vendoring this again from the web site but
it has not changed in many years. It may not be worth it. It may be simpler
just to do it by hand on occasion.

Builds and works fine.

Change-Id: I15c32c253287fb379459e1691cb9f5326d7ab6ca
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>