USB: get the rest of the files to compile.
The big goal here is that we'll have .h files in /sys/include and
library functions in /lib/libusb.a  so it's easy to build usb programs.

It all builds and as in all these conversion efforts I found errors
in the code.

The next step is probably to create /sys/include/usb/ and put
all .h files there, and make sure it all builds. Then we can
have /sys/src/cmds/usb/whatever and in there have example
usb programs.

The USB stuff was pretty incomplete so it's going to take a while
to get this rework right (assuming you all like it OK).

Change-Id: I2698cca6a369193cb7386c34146b45f964938793
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>