kernel: use segbrk instead of brk_
NOTE: this does not work. But maybe Ron can help to fix this.

Also,  Aki worked on this at

  https://github.com/Harvey-OS/harvey/commit/121208f9f0ea72e52c87021c4acb5d7ad51e9c37
  https://github.com/Harvey-OS/harvey/commit/5a029b6653ed1c1ff5daa6e66d8ac0fd3f7a88bc

In the first commit it cites Go, so maybe it is actually using segbrk
(that is not used anywhere else): thus maybe Go can be of help understanding
how to implement brk() and sbrk() in term of the current segbrk().

Change-Id: I9a269bf9fb4402167c2360e89d5605ad8972b34f
Signed-off-by: Giacomo Tesio <giacomo@tesio.it>