Sleep: partially undo the change of sleep to use awake (Fixes go runtime test hangs)
This CL restores the original sleep system call, and now the go runtime
test at least completes. It still has some bugs, but not the hang.

I am doing this because the awake-based sleep has a race condition that can cause
it to hang. The hang was seen in the go runtime tests, which would never complete.
Still worse, the runtime hang was really hard to diagnose, as it occured almost
randomly -- a classic sign of a race condition somewhere.

The race was in our sleep function in libc.

The proposed sleep replacement, using awake, looked like this:

void
sleep(int32_t millisecs)
{
	int64_t wakeup;
	char c;

	wakeup = awake(millisecs);

	while(!awakened(wakeup) && /*read(0, &c, 1) && */rendezvous(&wakeup, (void*)1) == (void*)~0)
		;
}

I've added the part in /* */

The code as written has a race that's not fixable in user mode.The test for exiting the
loop is two parts: awakened and rendezvous. It's possible for the awakened test to fail,
meaning there is time left; but is is further possible for
the awake to finish, and now the rendezvous will block. Note the commented-out-read: if you
uncomment it, and hit return at various times, the rendezvous will hang *depending on the value of the timeout
variable*. This is a classic race.

This broke the to Go runtime test, in various parts, in nonrepeatable ways.

I'm leaving most of the awake-based sleep there until we decide what to do with it. It's also a
warning to all of us to review global system call patches much more carefully. The libc sleep was
broken, and the breakage should have been caught in code review. It was not.

Finally, before we continue further global changes to system calls, we need a solid workload test.
I renew my call for a moratorium on system call changes until we can verify them with a workload.

Change-Id: If4eec8125048b03517cc759a1ddffbc48cf016a2
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>