kernel: remove awake (and related code), see #25
In b6622f5db2fe243a21f744e200f85de00e253fbe Ron expressed concerns about
the interaction between awake and rendezvous in the sleep libc function.
In particular, their designated usage as expressed in the libc version
of sleep(), lead to a race condition in Go tests when implemented in Go.

I cannot reproduce the issue in C, but given that all awake callers uses
the same approach, if sleep was broken, all the other functions are broken too.

Thus the appropriate things to do here is to remove the awake syscall
and all libc related functions before somebody starts using it and
until somebody is able to reproduce the issue in a C regression test
that uses awake and rendezvous properly.

Change-Id: Id89dc6a7a0cddabd190b7a90567ad96729faef05
Signed-off-by: Giacomo Tesio <giacomo@tesio.it>