newvga: remove references to iig from classic amd64cpu; create new type: amd64coreboot
The whole devvga stuff is a hopeless anachronism and I'd like to let it die at the
side of the road. It's looking like I can move forward if I write a new
screen.c so that's the plan to try to work this out:
devdraw -> cbscreen.c -> deviig.c

This is clearly lacking abstractions but one thing at a time.

The goal is to leave devdraw unchanged.

Change-Id: I4f18a5ff4c7bea8010569c70508e7480a72bde71
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>