vga: final realemu fix
realemu needs a pretty good type cleaning up. But this simple fix gets it to run.

OK, folks! You have a working realmodemem (maybe) and emulator (maybe).

Time to get this working now. Anyone? I have other things I hvae to do for a while,
but wanted to get this to a state that people could at least give it a try.

It's hard, sure, but not as hard as you might thing.

Change-Id: If6c8042d9e9306bb88eaf1ea59058d81b9969e09
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>