mmu: remove the check in vmap for machno != 0
I think this check was done because of the model: all vmaps
are done pre-smp-startup and none should be done after that, since
they're all copied. I'm just guessing. If so, it's a failure of vision,
since it fails badly for vga.

Anyway, currently, all cores share the same page tables, so it's not
an issue.

The way plan 9/harvey does page tables is really broken. We need someone
to fix it.

Change-Id: I9c56ecbc31c5d832a42d5d5bcaf683b8a8ffb496
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>