vga: don't open cga mem directly for screenconsole; use /dev/cons or #c/cons
Otherwise, once vga is turned on, there's no visible output.

Output looks like crap, now, but we can fix that. screencons should not direct
render, it turns out; it breaks everything once vga is turned on.

With this change, you are still able to intereact with rc on the console!

Change-Id: I6b5844d87e1a01aac10aa0af8a394eac79945bdb
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>