kernel: stop building system calls into klibc
and break the build, because amd64/qmalloc.c does an brk_ system call.

Time to stop everything any of us are doing and fix amd64/qmalloc.c, because
it's kernel code that does a brk_ system call and this is what was killing
the use of poolalloc in the vga code.

Told you that gdb would help :-)

So, yes, this breaks the kernel build, but the fix is to stop putting code in the
kernel that uses system calls. So Let's have this go through; that will encourage
people to fix the broken qmalloc.c

Change-Id: Ice015c549ba1b28fe2487e24e864f82da96edcab
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>