external pager: add a pager file to proc.
Later, if this is opened, and faults for a process fail, the info will be
written via qwrite, so that external processes can provide a page.

It's an experimental idea but would make mmap a user level entity and hence
it could be far more flexible than it is on unix.

Change-Id: I0c2bae80ad8118546ab0f5a1c818ac38997ed561
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>