ratrace: change output to make it easier to trace many procs
The old ratrace format was modeled after strace, and it got very confusing when many procs
were involved. It gets very hard to match system call call and return when things get interleaved
and It's made it very hard to debug Go issues.

This new format simplifies it and makes it very regular, meaning easy for programs
to trace the output, and hopefully will make Go debugging a lot easier.

The format is simple:
syscall entry:
E pid command syscallname [params ...]

syscall exit:
X pid command sysscallname [params ...] = result "errstr" startns stopns

Fork syscall exit when a process is created is the same as X,
but it outputs F, not X, and hence notifies tracing tools that this is a successful fork and they should look
at the return value and start tracing that pid.

This is going to make the Go debugging much easier. It's also easier to visually see the behavior.
It also made ratrace itself simpler.

I've used this format in other projects and it's really effective.

Change-Id: If983d7ada2c70655f6a6ec24bef96491492d01dd
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>