I2C: initial commit
i2c is a low performance bus which is everywhere now. We're going to need it, but there's
no good reason to jam it into the kernel. This is going to be a 9p i2c server. If it works
here we can also use it in akaros.

This code all comes from coreboot, and this initial commit is intended to preserve provenance.
A lot is going to change.

Change-Id: I531c7e7f900352daf9f79665ccc4fc7d6b5374d6
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>