boot: add a delay to before accessing the console
This is starting to look like a nasty always-been-there bug
we just found. Basically, if this 2.5 second sleep is placed
before the access of the server that we spawn, the weird null
pointer panic in PID 1 never happens.

It never happened anyway, save on qemu. This should allow us to
have smp qemu again.

Change-Id: I5f25c29ce820d08c2fd85819de473599e146a261
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>