Revert to known-good NIX scheduler from not-so-good AMP scheduler
In April 2012, right before the end of the joint NIX effort,
AMP scheduling went into NIX. It was kind of experimental and
we're not sure it really works.

This is a draft to revert to the known good scheduler from summer 2011.

This won't build, clearly. And some careful looking at sleep ought to be done.
But we don't care about AMP and there's not much point in keeping the AMP
scheduler in harvey.

I think a week of careful looking at this would be worth it. It would certainly
simplify our life.

This works!

Thanks Dan and everyone! The hangout was very useful!

This is a second try; please test. The first one i went a little too far on.

Change-Id: I9eecc60e7c5e8aefca73e6db0913f6f579451c56
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>