kernel: change relative include of io.h to absolute
The plan 9 authors did it as relative for a reason, I'm pretty sure,
but evidently we know better :-)

This is an example of changing a relative include to absolute.

Tested for full build/boot cycle.

IF YOU MAKE ANY MORE GLOBAL CHANGES OF THIS KIND, YOU MUST
ALSO TEST FOR A FULL BUILD/BOOT CYCLE.

Change-Id: Ic2538f6ee282df494974e3469995eb19a2c82578
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>