clang: we can now build all of userland and boot
usbd dies with a suicide, as it seems to try to use floating point
in a note handler, oops. I bet there is a memcpy and clang generates
code to use xmm.

I remove the call to build libz/regression.json because clang got upset
with it.

There are many warnings, still, but at least we can build. I am going to get the warnings
fixed tonight/tomorrow and then add clang to the travis test matrix.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>