Update bootstrap.sh
It now pulls in blaze.

We don't yet tell anyone to use blaze because we're still importing BUILD files, and
the one we have doesn't work. It appears to be for an older version.

./util/blaze //sys/src/libmach

2016/06/09 21:18:00 /home/rminnich/projects/harvey/harvey:6: While parsing cl were expecting [Func String LeftBrac EOF] but got Comma.
	in="elf64-x86-64",
 >^^^ < < <
2016/06/09 21:18:00 we couldn't find target //:sys/src/libmach

Our plan is to fix this one BUILD file so we can bootstrap others.

Here's the bigger question from Dan Cross: should we just be using Bazel?

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Author:    Sevki <s@sevki.org>