amd64/sdata.c: Make buf[] in atadrive() static.
The atadrive function was declaring a local 512 byte buf, i.e. a
stack variable. Big arrays like this on kernel stack are almost always
a bad idea. In this case we were getting a page fault on NULL while probing.

I'm pretty sure this function is single threaded, i.e. only one invocation
is ever active at a time. So I just made buf[] static. The sdata panics
promptly ceased.

If you ever want to make this function reentrant you'll need to use malloc to
allocate buf[] and be sure to free it on all the (many) exits.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>