Panic if there is no MP table
The code assumes, from the earliest stages, that apicbase
has been set. Right now, it's only set in the mp table parser.

The result is that there are obscure failures if no _MP_ is found.

We need to drag this code base into the third millenium but at least
for now make an error that is useful, instead of just a panic on
referencing address 0x20.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>